requirements.txt
alabaster==1.0.0
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
astroid==3.3.9
babel==2.17.0
beautifulsoup4==4.13.4
billiard==4.2.1
blis==1.3.0
bs4==0.0.2
catalogue==2.0.10
celery==5.5.3
certifi==2025.1.31
charset-normalizer==3.4.1
ChatterBot==1.2.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.21.1
cohesion==1.2.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
dill==0.3.9
distlib==0.3.9
Django==5.0.14
django-allauth==65.9.0
docutils==0.21.2
filelock==3.18.0
fillpdf==0.7.3
graphviz==0.20.3
greenlet==3.2.2
idna==3.10
imagesize==1.4.1
isort==6.0.1
Jinja2==3.1.6
kombu==5.5.4
langcodes==3.5.0
language_data==1.3.0
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mathparse==0.1.2
mccabe==0.7.0
mdurl==0.1.2
mssql-django==1.5
murmurhash==1.0.13
mysql-connector-python==9.3.0
numpy==2.2.6
oauthlib==3.2.2
packaging==24.2
pdf2image==1.17.0
pdfrw==0.4
pdfrw2==0.5.0
pillow==11.2.1
platformdirs==4.3.7
preshed==3.0.10
prompt_toolkit==3.0.51
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pylint==3.3.6
PyMuPDF==1.26.1
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
roman-numerals-py==3.1.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
snowballstemmer==2.2.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx==8.2.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.41
sqlparse==0.5.3
srsly==2.5.1
thinc==8.3.6
tomlkit==0.13.2
tqdm==4.67.1
twitter==1.19.6
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
virtualenv==20.30.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wrapt==1.17.2
gunicorn
psycopg2-binary
whitenoise